refactor: merge security code to SecurityConfiguration
